requirements.txt
flask
babel
humanize
python-slugify